requirements.txt
requests
requests[socks]
legacy-cgi
environs
python-telegram-bot==13.*